Autoliga_Botfile/requirements.txt
aiofiles==25.1.0
aiogram==3.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.1.4
coloredlogs==15.0.1
executor==23.2
fasteners==0.20
frozenlist==1.8.0
greenlet==3.3.1
humanfriendly==10.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.1
propcache==0.4.1
property-manager==3.0
psycopg==3.3.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
six==1.17.0
SQLAlchemy==2.0.46
typing-inspection==0.4.2
typing_extensions==4.15.0
verboselogs==1.7
yarl==1.22.0
